+++ src/server/shared/Utilities/ProgressBar.h
@@ -27,15 +27,16 @@ class barGoLink
     static char const * const empty;
     static char const * const full;
 
-    uint64 rec_no;
-    uint64 rec_pos;
-    uint64 num_rec;
-    uint64 indic_len;
+    int rec_no;
+    int rec_pos;
+    int num_rec;
+    int indic_len;
 
     public:
 
         void step( void );
-        barGoLink( uint64 );
+        barGoLink( int );
         ~barGoLink();
 };
 #endif
+
